+++ compiler/src/Language/Mimsa/Typechecker/Generalise.hs
@@ -3,22 +3,35 @@ module Language.Mimsa.Typechecker.Generalise
   )
 where
 
+import Data.List ((\\))
 import qualified Data.Map as M
+import qualified Data.Set as S
 import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.Typechecker
 
-removeLambdas :: MonoType -> [TypeIdentifier]
-removeLambdas (MTVar _ b) = [b]
-removeLambdas (MTFunction _ _t1 t2) = removeLambdas t2
-removeLambdas (MTPair _ a b) = removeLambdas a <> removeLambdas b
-removeLambdas (MTRecord _ as) = mconcat (removeLambdas . snd <$> M.toList as)
-removeLambdas (MTRecordRow _ as rest) =
-  mconcat (removeLambdas . snd <$> M.toList as)
-    <> removeLambdas rest
-removeLambdas (MTArray _ a) = removeLambdas a
-removeLambdas (MTPrim _ _) = mempty
-removeLambdas MTData {} = mempty
+freeTypeVars :: MonoType -> [TypeIdentifier]
+freeTypeVars (MTVar _ b) = [b]
+freeTypeVars (MTFunction _ t1 t2) = freeTypeVars t1 <> freeTypeVars t2
+freeTypeVars (MTPair _ a b) = freeTypeVars a <> freeTypeVars b
+freeTypeVars (MTRecord _ as) = mconcat (freeTypeVars . snd <$> M.toList as)
+freeTypeVars (MTRecordRow _ as rest) =
+  mconcat (freeTypeVars . snd <$> M.toList as)
+    <> freeTypeVars rest
+freeTypeVars (MTArray _ a) = freeTypeVars a
+freeTypeVars (MTPrim _ _) = mempty
+freeTypeVars (MTData _ _ as) = foldMap freeTypeVars as
 
-generalise :: MonoType -> Scheme
-generalise ty =
-  Scheme (removeLambdas ty) ty
+freeTypeVarsScheme :: Scheme -> [TypeIdentifier]
+freeTypeVarsScheme (Scheme vars t) =
+  freeTypeVars t \\ vars
+
+freeTypeVarsCtx :: Environment -> [TypeIdentifier]
+freeTypeVarsCtx (Environment env _ _) =
+  foldMap freeTypeVarsScheme (M.elems env)
+
+generalise :: Environment -> MonoType -> Scheme
+generalise env ty =
+  Scheme (S.toList $ S.fromList vars) ty
+  where
+    vars =
+      freeTypeVars ty \\ freeTypeVarsCtx env
